Update README.md
this is the project work for CourseEra as well as for Data Camp where I am analysizing the banking data and the churns rates of the banking customers why they are churning from the bank.
This project work helps to understand why customers are churning, what makes them to churn and how to reduce the churn rate.So basically this project work helps to understand the over all how  it can help banks to improve customer loyalty, increase revenue, and reduce costs. For this project following activities were done:
Collect and understand the data as well as scenario of the project.
Cleaning data, Statical analysis, Visualization and Clustering the data.